Fix HNAS error with unconfined filesystems
Unconfined filesystems created through CLI in HNAS caused driver's
parser to break. It was found that the command breaking could be
replaced by another, already used command, which also retrieves
filesystem information, and it is preferred, so previous command
has been replaced.

Additionally, administrator may not want to automatically mount
an unmounted filesystem, which may be unmounted for maintenance,
so added a flag to prevent driver from mounting it automatically.

Change-Id: I69e7a51ed4485783868c44e58b93bc69b58c8ac2
Closes-bug: #1571000