Fix Share Migration improper behavior for drivers
Tempest tests were not appropriate for driver-assisted migration,
so this was fixed.

Also, improved docstrings and fixed workflow for drivers when
implementing 2-phase migration to be accurate with tempest.

Driver-assisted migration now creates an additional
share instance to better handle and support driver methods.

Also, bug in update_access database queries prevented drivers
from maintaining a correct set of rules after migration, so this
was fixed.

Closes-bug: #1594922
Closes-bug: #1566815
Change-Id: If4bfaf7e9d963b83c13a6fea241c2eda14f7f409