Fix Share Migration not allowing access
Share migration code merged in newton intended to allow admins
to mount the share between phase1 and phase2. But API is
blocking it. This patch fixes the condition in API to allow
only in such scenario as originally intended.

Also, adds tempest test that validates such scenario.

To allow destination migrating share instance to be mounted
when driver-assisted migration approach is used, it is now
possible for drivers to return export locations update.

APIImpact

Closes-bug: #1619788

Change-Id: I76a7d8c3bdd597b951e700350f8f3f82bfb21e03